fix(symbol-graph): close Stage B review findings on PR #2313
Two independent Stage B gates (reviewer + test-engineer) both returned
NEEDS_REVISION on the first feedback commit, and converged on the same new
regression. Both findings were reproduced before being fixed.

1. The FB-007 guard was incomplete. It tested the full dotted path for the
   CURRENT root prefix only, and `''` is iterated first, so a namespace living
   under a later prefix never got a chance: `using App.Models;` with
   `src/App/Models/User.cs` present still returned an unrelated `App.cs`. That
   is the canonical .NET layout the same commit's JSDoc claimed was covered.
   The three probes now run as separate passes over every root prefix, most
   specific first, and the parent-as-file (nested type) probe is restricted to
   Java and Kotlin. A C# `using X.Y;` names a namespace, never a type; C#
   reaches a nested type through `using static` or an alias, neither
   distinguishable from a namespace import by the specifier alone. The cost is
   that a C# alias to a nested type no longer resolves, which is the right
   trade: a missing edge beats a false one.

2. The FB-011A masking fix introduced a regression of its own. A C# verbatim
   string ending in an escaped quote — @"say ""hi""" — contains a literal triple
   quote, so masking the triple-quote form first paired that tail with a later
   one and blanked every line between, deleting a real `namespace` declaration
   and dropping the boundary to the path fallback. That is worse than the bug it
   replaced. Masking now consumes the verbatim form first and is scoped by
   language.

Also from Stage B: two of the boundary tests were non-discriminating (the real
declaration preceded the literal, so they passed pre-fix) and are now labelled
as ordering companions rather than counted as proof; the mask-order regression,
the namespace-directory-of-subdirectories-only interaction, and several
edge cases now have pins; two JSDoc blocks that had detached from their
functions were re-attached; and the FB-001 fixture comment now states plainly
that its case alignment is enforced by the Linux CI shard, not by any local
assertion.

The dotted-module resolution tests moved to their own file to stay under the
500-line FR-006 cap.
